+++ packages/amplify-category-notifications/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-notifications@1.4.9...amplify-category-notifications@1.4.10) (2019-06-20)
+
+
+### Bug Fixes
+
+* **cli:** fix inquirer version ([#1690](https://github.com/aws-amplify/amplify-cli/issues/1690)) ([9246032](https://github.com/aws-amplify/amplify-cli/commit/9246032)), closes [#1688](https://github.com/aws-amplify/amplify-cli/issues/1688)
+
+
+
+
+
 ## [1.4.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-notifications@1.4.8...amplify-category-notifications@1.4.9) (2019-06-11)