+++ packages/graphql-function-transformer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/aws-amplify/amplify-cli/compare/graphql-function-transformer@1.1.1...graphql-function-transformer@1.1.2) (2019-06-26)
+
+**Note:** Version bump only for package graphql-function-transformer
+
+
+
+
+
 ## [1.1.1](https://github.com/aws-amplify/amplify-cli/compare/graphql-function-transformer@1.1.0...graphql-function-transformer@1.1.1) (2019-06-12)
 
 **Note:** Version bump only for package graphql-function-transformer