+++ packages/graphql-relational-schema-transformer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.10](https://github.com/aws-amplify/amplify-cli/compare/graphql-relational-schema-transformer@1.0.9...graphql-relational-schema-transformer@1.0.10) (2019-06-26)
+
+**Note:** Version bump only for package graphql-relational-schema-transformer
+
+
+
+
+
 ## [1.0.9](https://github.com/aws-amplify/amplify-cli/compare/graphql-relational-schema-transformer@1.0.8...graphql-relational-schema-transformer@1.0.9) (2019-06-12)
 
 **Note:** Version bump only for package graphql-relational-schema-transformer